backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aidemo-backend"
version = "0.1.0"
description = "FastAPI RAG Backend with LLM Integration"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.12.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "openai>=1.3.0",
    "boto3>=1.34.0",
    "azure-storage-blob>=12.19.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.1.0",
    "pypdf2>=3.0.0",
    "python-docx>=1.1.0",
    "unstructured>=0.11.0",
    "aiofiles>=23.2.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]